+++ CHANGELOG.md
@@ -1,4 +1,12 @@
-### HEAD
+###
+
+### 1.0.5 (March 9, 2021)
+
+  * Added the _render method to increase the robustness of $removeChild,
+  * Fixed an issue to prevent $append to fire twice the component method events for childs,
+  * Add The method $listenOnce,
+  * ...,
+
 
 ### 1.0.4 (February 18, 2021)